+++ lib/Service/AnyToPdf.php
@@ -40,17 +40,22 @@ class AnyToPdf
 {
   use \OCA\PdfDownloader\Traits\LoggerTrait;
 
+  const UNIVERSAL = '[universal]';
+  const FALLBACK = '[fallback]';
+  const PASS_THROUGH = '[pass-through]';
+
+  const DEFAULT_FALLBACK_CONVERTER = 'unoconv';
+
   /**
    * @var string Array of available converters per mime-type. These form a
    * chain. If part of the chain is again an error then the first succeeding
    * sub-converter wins.
    */
   const CONVERTERS = [
-    'message/rfc822' => [ 'mhonarc', [ 'wkhtmltopdf', 'unoconv', ], ],
+    'message/rfc822' => [ 'mhonarc', [ 'wkhtmltopdf', self::FALLBACK, ], ],
     'application/postscript' => [ 'ps2pdf', ],
     'image/tiff' => [ 'tiff2pdf' ],
-    'application/pdf' => [ 'passthrough' ],
-    'default' => [ 'unoconv', ],
+    'application/pdf' => [ self::PASS_THROUGH ],
   ];
 
   const DEFAULT_BLACKLIST = [
@@ -84,6 +89,15 @@ class AnyToPdf
   /** @var ExecutableFinder */
   protected $executableFinder;
 
+  /** @var string */
+  protected $fallbackConverter;
+
+  /** @var string */
+  protected $universalConverter;
+
+  /** @var bool */
+  protected $builtinConvertersDisabled;
+
   /**
    * @var array
    * Cache of found executables for the current request.
@@ -102,6 +116,88 @@ class AnyToPdf
     $this->executableFinder = $executableFinder;
     $this->logger = $logger;
     $this->l = $l;
+
+    $this->fallbackConverter = self::DEFAULT_FALLBACK_CONVERTER;
+  }
+
+  public function setFallbackConverter(?string $converter)
+  {
+    if (empty($converter)) {
+      $converter = self::DEFAULT_FALLBACK_CONVERTER;
+    }
+    $this->fallbackConverter = $converter;
+  }
+
+  public function getFallbackConverter():string
+  {
+    return $this->fallbackConverter;
+  }
+
+  public function setUniversalConverter(?string $converter)
+  {
+    $this->universalConverter = $converter;
+  }
+
+  public function getUniversalConverter():?string
+  {
+    return $this->universalConverter;
+  }
+
+  public function disableBuiltinConverters(bool $state = true)
+  {
+    $this->builtinConvertersDisabled = $state;
+  }
+
+  public function builtinConvertersDisabled():bool
+  {
+    return !empty($this->builtinConvertersDisabled);
+  }
+
+  /**
+   * Diagnose the state of the builtin-converter chains, i.e. try to find the
+   * binaries.
+   *
+   * @return array
+   */
+  public function findConverters()
+  {
+    $result = [];
+
+    if (!empty($this->universalConverter)) {
+      $executable = $this->executableFinder->find($this->universalConverter);
+      if (empty($executable)) {
+        $executable = $this->l->t('not found');
+      }
+      $result[self::UNIVERSAL] = [ [ $this->universalConverter => $executable ] ];
+    }
+    if ($this->builtinConvertersDisabled) {
+      return $result;
+    }
+    foreach (self::CONVERTERS as $mimeType => $converterChain) {
+      $result[$mimeType] = [];
+      foreach ($converterChain as $converters) {
+        if (!is_array($converters)) {
+          $converters = [ $converters ];
+        }
+        $probedConverters = [];
+        foreach ($converters as $converter) {
+          if ($converter == self::PASS_THROUGH) {
+            $probedConverters[$converter] = $this->l->t('pass through');
+            continue;
+          }
+          if ($converter == self::FALLBACK) {
+            $converter = $this->fallbackConverter;
+          }
+          $executable = $this->executableFinder->find($converter);
+          if (empty($executable)) {
+            $executable = $this->l->t('not found');
+          }
+          $probedConverters[$converter] = $executable;
+        }
+        $result[$mimeType][] = $probedConverters;
+      }
+    }
+    return $result;
   }
 
   /**
@@ -120,7 +216,19 @@ class AnyToPdf
       $mimeType = $this->mimeTypeDetector->detectString($data);
     }
 
-    $converters = self::CONVERTERS[$mimeType] ?? self::CONVERTERS['default'];
+    if (!empty($this->universalConverter)) {
+      try {
+        $data = $this->genericConvert($data, $mimeType, $this->universalConverter);
+      } catch (\Throwable $t) {
+        if ($this->builtinConvertersDisabled) {
+          throw new \RuntimeException($this->l->t('Universal converter "%1$s" fas failed trying to convert mime-type "%2$s"', [ $this->universalConverter, $mimeType ]));
+        } else {
+          $this->logException($t, 'Ignoring failed universal converter ' . $this->universalConverter);
+        }
+      }
+    }
+
+    $converters = self::CONVERTERS[$mimeType] ?? [ self::FALLBACK ];
 
     foreach ($converters as $converter) {
       if (!is_array($converter)) {
@@ -129,9 +237,16 @@ class AnyToPdf
 
       $convertedData = null;
       foreach  ($converter as $tryConverter) {
+        if ($tryConverter == self::FALLBACK) {
+          $tryConverter == $this->fallbackConverter;
+        }
         try {
           $method = $tryConverter . 'Convert';
-          $convertedData = $this->$method($data);
+          if (method_exists($this, $method)) {
+            $convertedData = $this->$method($data);
+          } else {
+            $convertedData = $this->genericConvert($data, $mimeType, $tryConverter);
+          }
           break;
         } catch (\Throwable $t) {
           $this->logException($t, 'Ignoring failed converter ' . $tryConverter);
@@ -150,6 +265,17 @@ class AnyToPdf
   protected function passthroughConvert(string $data):string
   {
     return $data;
+  }
+
+  protected function genericConvert(string $data, string $mimeType, string $converterName):string
+  {
+    $converter = $this->findExecutable($converterName);
+    $process = new Process([
+      $converter,
+      '--mime-type=' . $mimeType,
+    ]);
+    $process->setInput($data)->run();
+    return $process->getOutput();
   }
 
   protected function unoconvConvert(string $data):string